move capabilities munging into service
