+++ packages/wdio-electron-service/src/service.ts
@@ -1,48 +1,39 @@
-import { ChildProcessWithoutNullStreams, spawn } from 'child_process';
-import fs from 'fs-extra';
 import path from 'path';
-import split2 from 'split2';
 import { path as chromedriverPath } from 'chromedriver';
 import { remote } from 'webdriverio';
-import logger from '@wdio/logger';
-import tcpPortUsed from 'tcp-port-used';
-import EventEmitter from 'events';
 import { Capabilities, Options, Services, Frameworks } from '@wdio/types';
-import getFilePath from './getFilePath';
 import { SpectronClient } from '~/common/types';
-
-const log = logger('spectron');
-
-const POLL_INTERVAL = 100;
-const POLL_TIMEOUT = 10000;
-const DEFAULT_CONNECTION = {
-  protocol: 'http',
-  hostname: '127.0.0.1',
-  port: 9515,
-  path: '/',
-};
 
 function delay(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+function getBinaryPath(distPath: string, appName: string) {
+  const SupportedPlatform = {
+    darwin: 'darwin',
+    linux: 'linux',
+    win32: 'win32',
+  };
+
+  if (!Object.values(SupportedPlatform).includes(process.platform)) {
+    throw new Error(`Unsupported platform: ${process.platform}`);
+  }
+
+  const pathMap = {
+    darwin: `mac/${appName}.app/Contents/MacOS/${appName}`,
+    linux: `linux-unpacked/${appName}`,
+    win32: `win-unpacked/${appName}.exe`,
+  };
+
+  const electronPath = pathMap[process.platform as keyof typeof SupportedPlatform];
+
+  return `${distPath}/${electronPath}`;
 }
 
 export type ChromeOptions = {
   binary: string;
   args: string[];
   windowTypes: ['app', 'webview'];
-};
-
-export type Options = {
-  'port'?: number;
-  'path'?: string;
-  'protocol'?: string;
-  'hostname'?: string;
-  'outputDir'?: string;
-  'logFileName'?: string;
-  'chromedriverCustomPath'?: string;
-  'args'?: string[];
-  'browserName'?: 'chrome';
-  'goog:chromeOptions'?: ChromeOptions;
 };
 
 // export type Capabilities = {
@@ -54,29 +45,42 @@ export type Config = {
   outputDir: string;
 };
 
-function redirectLogStream(outputDir: string, process: ChildProcessWithoutNullStreams): void {
-  const logFile = getFilePath(outputDir, 'wdio-chromedriver.log');
-
-  // ensure file & directory exists
-  fs.ensureFileSync(logFile);
-
-  const logStream = fs.createWriteStream(logFile, { flags: 'w' });
-
-  process?.stdout.pipe(logStream);
-  process?.stderr.pipe(logStream);
-}
-
-export default class SpectronWorkerService implements Services.ServiceInstance {
+export default class ElectronWorkerService implements Services.ServiceInstance {
   constructor(
     options: Services.ServiceOption,
     caps: Capabilities.Capabilities,
     config: Omit<Options.Testrunner, 'capabilities'>,
   ) {
+    this.options = options;
+    const chromeArgs = [];
+    const isWin = process.platform === 'win32';
+
+    if (process.env.CI) {
+      chromeArgs.push('window-size=1280,800');
+      chromeArgs.push('blink-settings=imagesEnabled=false');
+      chromeArgs.push('enable-automation');
+      chromeArgs.push('disable-infobars');
+      chromeArgs.push('disable-extensions');
+      if (!isWin) {
+        // chromeArgs.push('headless'); - crashes on linux with xvfb
+        chromeArgs.push('no-sandbox');
+        chromeArgs.push('disable-gpu');
+        chromeArgs.push('disable-dev-shm-usage');
+        chromeArgs.push('disable-setuid-sandbox');
+        // chromeArgs.push('remote-debugging-port=9222');
+      }
+    }
+    caps.browserName = 'chrome';
+    caps['goog:chromeOptions'] = {
+      binary: getBinaryPath(options.appPath, options.appName),
+      args: chromeArgs,
+      windowTypes: ['app', 'webview'],
+    };
     this.wdOpts = {
-      protocol: options.protocol || DEFAULT_CONNECTION.protocol,
-      hostname: options.hostname || DEFAULT_CONNECTION.hostname,
-      port: options.port || DEFAULT_CONNECTION.port,
-      path: options.path || DEFAULT_CONNECTION.path,
+      // protocol: options.protocol || DEFAULT_CONNECTION.protocol,
+      // hostname: options.hostname || DEFAULT_CONNECTION.hostname,
+      // port: options.port || DEFAULT_CONNECTION.port,
+      // path: options.path || DEFAULT_CONNECTION.path,
       capabilities: caps,
     };
 
@@ -87,6 +91,8 @@ export default class SpectronWorkerService implements Services.ServiceInstance {
     console.log('constructor end', this.wdOpts);
   }
 
+  public options;
+
   public wdOpts;
 
   public browser?: SpectronClient;
@@ -94,8 +100,6 @@ export default class SpectronWorkerService implements Services.ServiceInstance {
   public outputDir;
 
   public chromedriverCustomPath;
-
-  public process: ChildProcessWithoutNullStreams | undefined;
 
   onWorkerStart(
     cid: string,
@@ -105,44 +109,55 @@ export default class SpectronWorkerService implements Services.ServiceInstance {
     execArgv: string[],
   ): void {
     console.log('zOMG onworkerStart', cid);
-    if (!this.process) {
-      void (async () => {
-        await this.startProcess();
-      })();
-      const exitHandler = () => {
-        void (async () => {
-          await this.killProcess.call(this);
-        })();
-      };
-      process.on('exit', exitHandler);
-      process.on('SIGINT', exitHandler);
-      process.on('uncaughtException', exitHandler);
-    }
   }
 
   beforeSession(
     config: Omit<Options.Testrunner, 'capabilities'>,
-    capabilities: Capabilities.RemoteCapability,
+    capabilities: Capabilities.Capabilities,
     specs: string[],
     cid: string,
   ): void {
     console.log('beforesession yo');
+    const chromeArgs = [];
+    const isWin = process.platform === 'win32';
+
+    if (process.env.CI) {
+      chromeArgs.push('window-size=1280,800');
+      chromeArgs.push('blink-settings=imagesEnabled=false');
+      chromeArgs.push('enable-automation');
+      chromeArgs.push('disable-infobars');
+      chromeArgs.push('disable-extensions');
+      if (!isWin) {
+        // chromeArgs.push('headless'); - crashes on linux with xvfb
+        chromeArgs.push('no-sandbox');
+        chromeArgs.push('disable-gpu');
+        chromeArgs.push('disable-dev-shm-usage');
+        chromeArgs.push('disable-setuid-sandbox');
+        // chromeArgs.push('remote-debugging-port=9222');
+      }
+    }
+
+    if (process.env.SPECTRON_APP_ARGS) {
+      chromeArgs.push(...process.env.SPECTRON_APP_ARGS.split(','));
+    }
+
+    const { appPath, appName } = this.options;
+
+    capabilities.browserName = 'chrome';
+    capabilities['goog:chromeOptions'] = {
+      binary: getBinaryPath(appPath, appName),
+      args: chromeArgs,
+      windowTypes: ['app', 'webview'],
+    };
+
     // set capabilities
     // this.browser = browser;
   }
 
   async afterTest(test: Frameworks.Test, context: never, results: Frameworks.TestResult): Promise<void> {
-    // if (context === 'afterEach') {
-    // quit app
-    // console.log('zOMG afterTest', test, context, results);
-    // void (async () => {
     await this.browser?.exitElectronApp();
     await delay(1000);
-    await this.killProcess();
-    await this.startProcess();
     await remote(this.wdOpts);
-    // })();
-    // }
   }
 
   async after(result: number, capabilities: Capabilities.RemoteCapability, specs: string[]): Promise<void> {
@@ -156,40 +171,5 @@ export default class SpectronWorkerService implements Services.ServiceInstance {
     results: any,
   ): Promise<void> {
     console.log('complete');
-  }
-
-  async killProcess(): Promise<void> {
-    if (this.process) {
-      console.log('killing process on port', this.process.pid);
-      await this.process.kill();
-      this.process = undefined;
-      // await process.kill(-(this.process.pid as number));
-    }
-  }
-
-  async startProcess(): Promise<void> {
-    console.log('zOMG startProcess');
-    const port = this.wdOpts.port as number;
-    const args = [`--port=${port}`, `--url-base=${this.wdOpts.path as string}`];
-    let command = this.chromedriverCustomPath;
-    log.info(`Start Chromedriver (${command}) with args ${args.join(' ')}`);
-    if (!fs.existsSync(command)) {
-      log.warn('Could not find chromedriver in default path: ', command);
-      log.warn('Falling back to use global chromedriver bin');
-      command = process?.platform === 'win32' ? 'chromedriver.exe' : 'chromedriver';
-    }
-    const chromedriverProcess = spawn(command, args, { detached: true });
-    console.log('spawned CD process', chromedriverProcess.pid);
-    this.process = chromedriverProcess;
-
-    if (this.outputDir && typeof this.outputDir === 'string') {
-      redirectLogStream(this.outputDir, chromedriverProcess);
-    } else {
-      const split = split2();
-      (chromedriverProcess.stdout.pipe(split) as EventEmitter).on('data', (...logArgs) => log.info(...logArgs));
-      (chromedriverProcess.stderr.pipe(split) as EventEmitter).on('data', (...logArgs) => log.warn(...logArgs));
-    }
-
-    await tcpPortUsed.waitUntilUsed(port, POLL_INTERVAL, POLL_TIMEOUT);
   }
 }
